Split apart asset loading module
